+++ diesel/src/types/impls/floats.rs
@@ -8,31 +8,31 @@ use self::byteorder::{ReadBytesExt, WriteBytesExt, BigEndian};
 use types::{self, FromSql, ToSql, IsNull};
 
 impl<DB: Backend<RawValue=[u8]>> FromSql<types::Float, DB> for f32 {
-    fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error>> {
+    fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {
         let mut bytes = not_none!(bytes);
-        bytes.read_f32::<BigEndian>().map_err(|e| Box::new(e) as Box<Error>)
+        bytes.read_f32::<BigEndian>().map_err(|e| Box::new(e) as Box<Error+Send+Sync>)
     }
 }
 
 impl<DB: Backend> ToSql<types::Float, DB> for f32 {
-    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error>> {
+    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         out.write_f32::<BigEndian>(*self)
             .map(|_| IsNull::No)
-            .map_err(|e| Box::new(e) as Box<Error>)
+            .map_err(|e| Box::new(e) as Box<Error+Send+Sync>)
     }
 }
 
 impl<DB: Backend<RawValue=[u8]>> FromSql<types::Double, DB> for f64 {
-    fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error>> {
+    fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {
         let mut bytes = not_none!(bytes);
-        bytes.read_f64::<BigEndian>().map_err(|e| Box::new(e) as Box<Error>)
+        bytes.read_f64::<BigEndian>().map_err(|e| Box::new(e) as Box<Error+Send+Sync>)
     }
 }
 
 impl<DB: Backend> ToSql<types::Double, DB> for f64 {
-    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error>> {
+    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         out.write_f64::<BigEndian>(*self)
             .map(|_| IsNull::No)
-            .map_err(|e| Box::new(e) as Box<Error>)
+            .map_err(|e| Box::new(e) as Box<Error+Send+Sync>)
     }
 }