+++ diesel/src/types/impls/primitives.rs
@@ -32,9 +32,9 @@ expression_impls! {
 impl NotNull for () {}
 
 impl<DB: Backend<RawValue=[u8]>> FromSql<types::VarChar, DB> for String {
-    fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error>> {
+    fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {
         let bytes = not_none!(bytes);
-        String::from_utf8(bytes.into()).map_err(|e| Box::new(e) as Box<Error>)
+        String::from_utf8(bytes.into()).map_err(|e| Box::new(e) as Box<Error+Send+Sync>)
     }
 }
 
@@ -42,16 +42,16 @@ impl<DB> ToSql<types::VarChar, DB> for String where
     DB: Backend,
     for<'a> &'a str: ToSql<types::VarChar, DB>,
 {
-    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error>> {
+    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         (&self as &str).to_sql(out)
     }
 }
 
 impl<'a, DB: Backend> ToSql<types::VarChar, DB> for &'a str {
-    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error>> {
+    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         out.write_all(self.as_bytes())
             .map(|_| IsNull::No)
-            .map_err(|e| Box::new(e) as Box<Error>)
+            .map_err(|e| Box::new(e) as Box<Error+Send+Sync>)
     }
 }
 
@@ -59,7 +59,7 @@ impl<DB> FromSql<types::Text, DB> for String where
     DB: Backend,
     String: FromSql<types::VarChar, DB>,
 {
-    fn from_sql(bytes: Option<&DB::RawValue>) -> Result<Self, Box<Error>> {
+    fn from_sql(bytes: Option<&DB::RawValue>) -> Result<Self, Box<Error+Send+Sync>> {
         <Self as FromSql<types::VarChar, DB>>::from_sql(bytes)
     }
 }
@@ -68,7 +68,7 @@ impl<DB> ToSql<types::Text, DB> for String where
     DB: Backend,
     for<'a> &'a str: ToSql<types::Text, DB>,
 {
-    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error>> {
+    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         (&self as &str).to_sql(out)
     }
 }
@@ -77,13 +77,13 @@ impl<'a, DB> ToSql<types::Text, DB> for &'a str where
     DB: Backend,
     &'a str: ToSql<types::VarChar, DB>,
 {
-    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error>> {
+    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         ToSql::<types::VarChar, DB>::to_sql(self, out)
     }
 }
 
 impl<DB: Backend<RawValue=[u8]>> FromSql<types::Binary, DB> for Vec<u8> {
-    fn from_sql(bytes: Option<&DB::RawValue>) -> Result<Self, Box<Error>> {
+    fn from_sql(bytes: Option<&DB::RawValue>) -> Result<Self, Box<Error+Send+Sync>> {
         Ok(not_none!(bytes).into())
     }
 }
@@ -92,16 +92,16 @@ impl<DB> ToSql<types::Binary, DB> for Vec<u8> where
     DB: Backend,
     for<'a> &'a [u8]: ToSql<types::Binary, DB>,
 {
-    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error>> {
+    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         (&self as &[u8]).to_sql(out)
     }
 }
 
 impl<'a, DB: Backend> ToSql<types::Binary, DB> for &'a [u8] {
-    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error>> {
+    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         out.write_all(self)
             .map(|_| IsNull::No)
-            .map_err(|e| Box::new(e) as Box<Error>)
+            .map_err(|e| Box::new(e) as Box<Error+Send+Sync>)
     }
 }
 
@@ -111,7 +111,7 @@ impl<'a, T: ?Sized, ST, DB> ToSql<ST, DB> for Cow<'a, T> where
     DB: Backend + HasSqlType<ST>,
     T::Owned: ToSql<ST, DB>,
 {
-    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error>> {
+    fn to_sql<W: Write>(&self, out: &mut W) -> Result<IsNull, Box<Error+Send+Sync>> {
         match self {
             &Cow::Borrowed(ref t) => t.to_sql(out),
             &Cow::Owned(ref t) => t.to_sql(out),
@@ -124,7 +124,7 @@ impl<'a, T: ?Sized, ST, DB> FromSql<ST, DB> for Cow<'a, T> where
     DB: Backend + HasSqlType<ST>,
     T::Owned: FromSql<ST, DB>,
 {
-    fn from_sql(bytes: Option<&DB::RawValue>) -> Result<Self, Box<Error>> {
+    fn from_sql(bytes: Option<&DB::RawValue>) -> Result<Self, Box<Error+Send+Sync>> {
         T::Owned::from_sql(bytes).map(Cow::Owned)
     }
 }
@@ -134,7 +134,7 @@ impl <'a, T: ?Sized, ST, DB> ::types::FromSqlRow<ST, DB> for Cow<'a, T> where
     DB: Backend + HasSqlType<ST>,
     Cow<'a, T>: FromSql<ST, DB>,
 {
-    fn build_from_row<R: ::row::Row<DB>>(row: &mut R) -> Result<Self, Box<Error>> {
+    fn build_from_row<R: ::row::Row<DB>>(row: &mut R) -> Result<Self, Box<Error+Send+Sync>> {
         FromSql::<ST, DB>::from_sql(row.take())
     }
 }