+++ diesel_tests/tests/types.rs
@@ -372,13 +372,13 @@ fn third_party_crates_can_add_new_types() {
     }
 
     impl FromSql<MyInt, Pg> for i32 {
-        fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error>> {
+        fn from_sql(bytes: Option<&[u8]>) -> Result<Self, Box<Error+Send+Sync>> {
             FromSql::<Integer, Pg>::from_sql(bytes)
         }
     }
 
     impl FromSqlRow<MyInt, Pg> for i32 {
-        fn build_from_row<R: ::diesel::row::Row<Pg>>(row: &mut R) -> Result<Self, Box<Error>> {
+        fn build_from_row<R: ::diesel::row::Row<Pg>>(row: &mut R) -> Result<Self, Box<Error+Send+Sync>> {
             FromSql::<MyInt, Pg>::from_sql(row.take())
         }
     }